requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
yfinance==1.2.0
openai==1.51.0
feedparser==6.0.11
apscheduler==3.10.4
sqlalchemy==2.0.35
aiosqlite==0.20.0
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.9.2
curl_cffi==0.13.0
